+++ apps/router-e2e/app.config.js
@@ -60,6 +60,8 @@ module.exports = {
               : process.env.E2E_ROUTER_ASYNC || false,
         root: path.join('__e2e__', process.env.E2E_ROUTER_SRC ?? 'static-rendering', 'app'),
         origin: 'http://localhost:3000/',
+        sitemap:
+          process.env.E2E_ROUTER_SITEMAP === 'false' ? false : process.env.E2E_ROUTER_SITEMAP,
       },
     ],
   ],