+++ packages/expo-router/plugin/src/index.ts
@@ -33,6 +33,8 @@ const withRouter: ConfigPlugin<
     root?: string;
     /** Should Async Routes be enabled, currently only `development` is supported. */
     asyncRoutes?: string | { android?: string; ios?: string; web?: string; default?: string };
+    /** Should the sitemap be generated. Defaults to `true` */
+    sitemap?: boolean;
   } | void
 > = (config, _props) => {
   const props = _props || {};